Extract dispatcher handlers to separate files
