+++ config.py
@@ -1,2 +1,3 @@
 TOKEN = "YOUR_TOKEN"
+INVITE_CODE = ''
 BASE_URL = "https://habr.com"
